Make 'Home Home' shortcut always open capital's city screen
